Worked with team, figured out the method for adding and deleting walls. Some small changes about previous functions has been made. Wait for communicating with other groups to do the merging part.
