--- .idea/libraries/Maven__org_apiguardian_apiguardian_api_1_1_2.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: org.apiguardian:apiguardian-api:1.1.2">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/apiguardian/apiguardian-api/1.1.2/apiguardian-api-1.1.2.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/org/apiguardian/apiguardian-api/1.1.2/apiguardian-api-1.1.2-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/apiguardian/apiguardian-api/1.1.2/apiguardian-api-1.1.2-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
