--- .idea/libraries/Maven__org_openjfx_javafx_controls_19.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: org.openjfx:javafx-controls:19">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/openjfx/javafx-controls/19/javafx-controls-19.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/org/openjfx/javafx-controls/19/javafx-controls-19-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/openjfx/javafx-controls/19/javafx-controls-19-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
