--- .idea/libraries/lib.xml
@@ -1,14 +1,0 @@
-<component name="libraryTable">
-  <library name="lib">
-    <CLASSES>
-      <root url="file://$PATH_TO_FX$" />
-      <root url="file://$PATH_TO_FX$" />
-      <root url="file://$PATH_TO_FX$" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-    <jarDirectory url="file://$PATH_TO_FX$" recursive="false" />
-    <jarDirectory url="file://$PATH_TO_FX$" recursive="false" />
-    <jarDirectory url="file://$PATH_TO_FX$" recursive="false" />
-  </library>
-</component>
